pyproject.toml
[project]
name = "transcribe-one-video"
version = "0.1.0"
requires-python = ">= 3.12.3"
dependencies = [
    "pyannote-audio>=4.0.1",
    "python-dotenv>=1.1.1",
]
